requirements.txt
asgiref==3.3.4
Django==3.2
pytz==2021.1
sqlparse==0.4.1
requests==2.25.1
aiohttp==3.7.4.post0
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
gunicorn==20.0.4
